Fix: disable delete/edit bookings for next day and previous

+++ booking/views.py
@@ -3,8 +3,23 @@ from .models import Booking
 from django.contrib import messages
 from .forms import BookingForm
 from django.contrib.auth.decorators import login_required
-import datetime
+from datetime import timedelta
 # Create your views here.
+
+# Checks if a specified date is in the future or not
+def is_date_in_future(date):
+    """
+    Checks if a passed in `date` is in the future
+
+    ``Context``
+
+        date passed in as a parameter to check if it is in the future
+    """
+    today = date.today() + timedelta(days=1)
+    return date > today
+
+
+
 @login_required
 def booking(request):
     """
@@ -43,19 +58,6 @@ def booking(request):
     )
 
 
-# Checks if a specified date is in the future or not
-def is_date_in_future(date):
-    """
-    Checks if a passed in `date` is in the future
-
-    ``Context``
-
-        date passed in as a parameter to check if it is in the future
-    """
-    today = date.today()
-    return date > today
-
-
 @login_required
 def booking_edit(request, id):
     """
@@ -77,8 +79,6 @@ def booking_edit(request, id):
     booking = get_object_or_404(queryset, id=id)
     booking_form = BookingForm(data=request.POST, instance=booking)
     booking_date = booking.booking_date
-    print(booking_date)
-    print(is_date_in_future(booking_date))
     if request.method == "POST"  and is_date_in_future(booking_date):
         if booking_form.is_valid() and booking.user == request.user:
             booking = booking_form.save(commit=False)
@@ -102,7 +102,7 @@ def booking_edit(request, id):
             }
         )
     else: 
-        messages.add_message(request, messages.ERROR, "You cannot edit a booking which is in the past")
+        messages.add_message(request, messages.ERROR, "You cannot edit a booking which is today, tomorrow or in the past")
         return redirect("my_profile")
 
 @login_required
@@ -117,16 +117,17 @@ def booking_delete(request, id):
     """
     queryset = Booking.objects.all()
     booking = get_object_or_404(queryset, id=id)
+    booking_date = booking.booking_date
 
-    if booking.user == request.user:
+    if booking.user == request.user and is_date_in_future(booking_date):
         booking.delete()
         messages.add_message(request, messages.SUCCESS, "Booking was deleted!")
-    else:
-        messages.add_message(
-            request, messages.ERROR, "You can only delete your own bookings!"
-        )
+    if is_date_in_future(booking_date):
+        return redirect("home_urls")
+    else: 
+        messages.add_message(request, messages.ERROR, "You cannot delete a booking which is today, tomorrow or in the past")
+        return redirect("my_profile")
 
-    return redirect("home_urls")
 
 
 def trigger_login_message(request):
